Исправит компонент отсутствия предложений.
