+++ src/components/no-city-offers.tsx
@@ -4,9 +4,14 @@ type NoCityOffersProps = {
 
 export default function NoCityOffers ({cityName}: NoCityOffersProps): JSX.Element {
   return (
-    <>
-      <b className="cities__status">No places to stay available</b>
-      <p className="cities__status-description">We could not find any property available at the moment in {cityName}</p>
-    </>
+    <div className="cities__places-container cities__places-container--empty container">
+      <section className="cities__no-places">
+        <div className="cities__status-wrapper tabs__content">
+          <b className="cities__status">No places to stay available</b>
+          <p className="cities__status-description">We could not find any property available at the moment in {cityName}</p>
+        </div>
+      </section>
+      <div className="cities__right-section"></div>
+    </div>
   );
 }
